+++ lib/msal-browser/src/index.ts
@@ -1,2 +1,15 @@
-// App
+// App and config
 export { PublicClientApplication } from "./app/PublicClientApplication";
+export { Configuration } from "./app/Configuration";
+
+// Common Object Formats
+export {
+    // Request
+    AuthenticationParameters,
+    TokenExchangeParameters,
+    // Response
+    AuthResponse,
+    // Error
+    AuthError,
+    AuthErrorMessage
+} from "msal-common";
