+++ lib/msal-common/src/app/module/AuthorizationCodeModule.ts
@@ -5,10 +5,14 @@
 
 // inheritance
 import { AuthModule } from "./AuthModule";
-import { MsalPublicClientConfiguration } from "../MsalPublicClientConfiguration";
+// app
+import { MsalPublicClientConfiguration } from "../config/MsalPublicClientConfiguration";
+// request
 import { AuthenticationParameters } from "../../request/AuthenticationParameters";
 import { TokenExchangeParameters } from "../../request/TokenExchangeParameters";
+// response
 import { TokenResponse } from "../../response/TokenResponse";
+import { ClientConfigurationError } from "../../error/ClientConfigurationError";
 
 /**
  * AuthorizationCodeModule class
@@ -17,9 +21,17 @@ import { TokenResponse } from "../../response/TokenResponse";
  * 
  */
 export class AuthorizationCodeModule extends AuthModule {
+
+    // Application config
+    protected config: MsalPublicClientConfiguration;
     
     constructor(configuration: MsalPublicClientConfiguration) {
-        super(configuration);
+        super({
+            storageInterface: configuration.storageInterface,
+            networkInterface: configuration.networkInterface,
+            cryptoInterface: configuration.cryptoInterface
+        });
+        this.config = configuration;
     }
 
     async createLoginUrl(request: AuthenticationParameters): Promise<string> {
@@ -33,4 +45,43 @@ export class AuthorizationCodeModule extends AuthModule {
     async acquireToken(request: TokenExchangeParameters): Promise<TokenResponse> {
         throw new Error("Method not implemented.");
     }
+
+    // #region Getters and setters
+
+    /**
+     *
+     * Use to get the redirect uri configured in MSAL or null.
+     * Evaluates redirectUri if its a function, otherwise simply returns its value.
+     * @returns {string} redirect URL
+     *
+     */
+    public getRedirectUri(): string {
+        if (this.config.auth.redirectUri) {
+            if (typeof this.config.auth.redirectUri === "function") {
+                return this.config.auth.redirectUri();
+            }
+            return this.config.auth.redirectUri;
+        } else {
+            throw ClientConfigurationError.createRedirectUriEmptyError();
+        }
+    }
+
+    /**
+     * Use to get the post logout redirect uri configured in MSAL or null.
+     * Evaluates postLogoutredirectUri if its a function, otherwise simply returns its value.
+     *
+     * @returns {string} post logout redirect URL
+     */
+    public getPostLogoutRedirectUri(): string {
+        if (this.config.auth.postLogoutRedirectUri) {
+            if (typeof this.config.auth.postLogoutRedirectUri === "function") {
+                return this.config.auth.postLogoutRedirectUri();
+            }
+            return this.config.auth.postLogoutRedirectUri;
+        } else {
+            throw ClientConfigurationError.createPostLogoutRedirectUriEmptyError();
+        }
+    }
+
+    // #endregion
 }
