+++ lib/msal-common/src/cache/ICacheStorage.ts
@@ -3,11 +3,37 @@
  * Licensed under the MIT License.
  */
 
+/**
+ * Interface class which implement cache storage functions used by MSAL to perform validity checks, and store tokens.
+ */
 export interface ICacheStorage {
-    setItem(key: string, value: string): void,
-    getItem(key: string): string,
-    removeItem(key: string): void,
-    containsKey(key: string): boolean,
-    getKeys(): string[],
-    clear(): void
+    /**
+     * Function to set item in cache.
+     * @param key 
+     * @param value 
+     */
+    setItem(key: string, value: string): void;
+    /**
+     * Function which retrieves item from cache.
+     * @param key 
+     */
+    getItem(key: string): string;
+    /**
+     * Function which removes item from cache.
+     * @param key 
+     */
+    removeItem(key: string): void;
+    /**
+     * Function which returns boolean whether cache contains a specific key.
+     * @param key 
+     */
+    containsKey(key: string): boolean;
+    /**
+     * Function which retrieves all current keys from the cache.
+     */
+    getKeys(): string[];
+    /**
+     * Function which clears cache.
+     */
+    clear(): void;
 }
