+++ lib/msal-common/src/response/AuthResponse.ts
@@ -3,12 +3,23 @@
  * Licensed under the MIT License.
  */
 
+/**
+ * AuthResponse base type returned by MSAL library on success
+ * - state: User given state
+ * - scopes: consented scopes
+ */
 export type AuthResponse = {
+    scopes: Array<string>;
     state: string;
 };
 
+/**
+ * Builds a response that only sets state
+ * @param responseState 
+ */
 export function buildResponseStateOnly(responseState: string) : AuthResponse {
     return {
+        scopes: null,
         state: responseState
     };
 }
