+++ lib/msal-common/src/utils/crypto/ICrypto.ts
@@ -17,7 +17,18 @@ export type PKCECodes = {
  * Interface for crypto functions used by library
  */
 export interface ICrypto {
+    /**
+     * base64 Encode string
+     * @param input 
+     */
     base64Encode(input: string): string;
+    /**
+     * base64 decode string
+     * @param input 
+     */
     base64Decode(input: string): string;
+    /**
+     * Generate PKCE codes for OAuth. See RFC here: https://tools.ietf.org/html/rfc7636
+     */
     generatePKCECodes(): Promise<PKCECodes>;
 }
